Split Nodestate controller into ebpfSyncer
Split the IngressNodeFirewallNodeState controller. Move the eBPF logic
into a thread safe Singleton and decouple the reconciler from the
eBPF programming.
eBPF logic will be managed by ebpfSyncer as a single point of contact
that all reconciliation runs will send their desired state of
interface rules to. On the other side, ebpfSyncer makes sure that
rules are attached and detached from / to the host's interfaces.
Another advantage of creating such an abstraction is that both the
reconciler and the eBPF logic can be unit tested independently.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>