Implement libxdploader stub for non-cgo compilation
Both the daemon and the main operator are in the same module.
Therefore, the go compiler will build the libxdploader even for the main
operator. In order to avoid this dependency, create a stub libxdploader
that will be used for non-CGO compilation, and use build tags to compile
the C dependencies only when CGO is used.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>
(cherry picked from commit bbe26e73d0073c73a1cc29a4978046c574220c5d)