kind: Mount bpffs individually into each docker container
Now that we are not mounting the fs any more as part of the application
logic, make sure that we mount it during kind setup. The easiest way to
achieve that is by using a DaemonSet that will run on each host and
mount bpffs during pod startup with an initContainer.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>