Daemon: ensure we know about all managed interfaces
If the daemon is restarted, we start not knowing about any
of the managed interfaces. If we perform activities like
reseting everything due to IngressNodeFirewallNodeState
deletion for example, we will not cleanup all XDP programs.
IngNodeFwController knows all pinned interfaces and ensure
we know about them in ebpfSingleton.

Future work is needed to refactor this to not have two
seperate sources of information about the same thing.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>